+++ Tests.Common/UnitTests.cs
@@ -40,6 +40,7 @@ using Rdmp.Core.DataLoad.Modules.DataFlowOperations;
 using Rdmp.Core.MapsDirectlyToDatabaseTable;
 using Rdmp.Core.Repositories;
 using Rdmp.Core.ReusableLibraryCode.Checks;
+using Rdmp.Core.Setting;
 
 namespace Tests.Common;
 
@@ -577,6 +578,10 @@ public class UnitTests
             var lmd = WhenIHaveA<LoadMetadata>(repository);
 
             return (T)(object)new LoadMetadataCatalogueLinkage(repository, lmd,cata);
+        }
+
+        if (typeof(T) == typeof(Setting)) {
+            return (T)(object)new Setting(repository.CatalogueRepository, "", "");
         }
 
         throw new TestCaseNotWrittenYetException(typeof(T));